+++ da_vinci/da_vinci/core/json.py
@@ -5,6 +5,8 @@ JSON Utility functionality
 from datetime import datetime
 from json import JSONEncoder
 
+from da_vinci.core.immutable_object import ObjectBody
+
 
 class DateTimeEncoder(JSONEncoder):
     '''
@@ -13,4 +15,18 @@ class DateTimeEncoder(JSONEncoder):
     def default(self, obj):
         if isinstance(obj, datetime):
             return obj.isoformat()
+        return JSONEncoder.default(self, obj)
+
+
+class DaVinciObjectEncoder(JSONEncoder):
+    '''
+    JSONEncoder class that encodes commonly used framework objects
+    '''
+    def default(self, obj):
+        if isinstance(obj, ObjectBody):
+            return obj.to_dict()
+
+        if isinstance(obj, datetime):
+            return obj.isoformat()
+
         return JSONEncoder.default(self, obj)